Created Audio Manager, added prefab to title screen
Created an Audio Manager singleton called AudioManager and made a prefab of it. I've put an instance of the prefab in the title screen. It doesn't destroy on load, so it should persist through scene changes.

Its functionality can be called in two ways:
AudioManager.playMusic([AudioClip Here]);
AudioManager.playSFX([AudioClip Here]);